Refactor indexer api (#352)
* feat: first pass at splitting up api functionality

* chore: cleanup api entrypoint file, add utils file with helper methods

* chore: standardize commands, move from npm -> yarn commands, nits

* chore: update changelog

* fix: update Dockerfile with new prisma:generate command
