+++ packages/indexer/docker-compose.rpi.yml
@@ -4,7 +4,7 @@ services:
     image: kekonen/indexer:latest
     build: .
     restart: always
-    command: npm run start indexer
+    command: yarn start:core
     env_file: .env
     environment:
       - PROGRAM=core
@@ -17,7 +17,7 @@ services:
     image: kekonen/indexer:latest
     build: .
     restart: always
-    command: npm run start indexer
+    command: yarn start:api
     ports:
       - "8080:9090"
       - "8081:3000"
