+++ packages/indexer/tsconfig.json
@@ -1,5 +1,4 @@
 {
-  
   "compilerOptions": {
     "outDir": "./dist",
     "rootDir": "./src",
